C4DIC Patient with Minimization
* Introduces a data minized version of Patient inspired by
shc-vaccine-credential-ig, the purpose of which is to have add
a minimal number of MS items and to change the cardinality of
others so that only information that is stored on an insurance
card is actually represented.

* Changes gender cardinality to 0..0 because of group feedback,
and per above direction.

* Changes photo cardinality to 0..0 because attachments would
be too big to ever store this in alternate formats

* Introduces an invariant called c4dic-name-invariant, also
inspired by shc-vaccine-credential-ig to allow for flexiblity
on name.

* Changes the inherited description from Patient because it
includes animals...

* Added a comment to birthDate and suggestions